Add support for optional MultiAddress.
