+++ rust/chains/tw_polkadot/src/signer.rs
@@ -2,6 +2,7 @@
 //
 // Copyright © 2017 Trust Wallet.
 
+use crate::ctx_from_tw;
 use crate::tx_builder::TxBuilder;
 use tw_coin_entry::coin_context::CoinContext;
 use tw_coin_entry::error::prelude::*;
@@ -9,12 +10,8 @@ use tw_coin_entry::signing_output_error;
 use tw_keypair::{ed25519::sha512::KeyPair, traits::KeyPairTrait};
 use tw_proto::Polkadot::Proto;
 use tw_scale::ToScale;
-use tw_ss58_address::{NetworkId, SS58Address};
+use tw_ss58_address::SS58Address;
 use tw_substrate::*;
-
-pub fn network_id_from_tw(input: &'_ Proto::SigningInput<'_>) -> EncodeResult<NetworkId> {
-    Ok(NetworkId::try_from(input.network as u16).map_err(|_| EncodeError::InvalidNetworkId)?)
-}
 
 pub struct PolkadotSigner;
 
@@ -32,10 +29,12 @@ impl PolkadotSigner {
         input: Proto::SigningInput<'_>,
     ) -> SigningResult<Proto::SigningOutput<'static>> {
         let key_pair = KeyPair::try_from(input.private_key.as_ref())?;
-        let network_id = network_id_from_tw(&input)?;
-        let account =
-            SubstrateAddress(SS58Address::from_public_key(key_pair.public(), network_id)?);
-        let unsigned_tx = TxBuilder::unsigned_tx_from_proto(coin, &input)?;
+        let ctx = ctx_from_tw(&input)?;
+        let account = ctx.multi_address(SubstrateAddress(SS58Address::from_public_key(
+            key_pair.public(),
+            ctx.network,
+        )?));
+        let unsigned_tx = TxBuilder::unsigned_tx_from_proto(&ctx, coin, &input)?;
         let signed_tx = unsigned_tx.sign(account, &key_pair)?;
         let encoded = signed_tx.to_scale();
